extra/rpi-mqtt/requirements.txt
paho-mqtt
RPI.GPIO
adafruit-blinka
adafruit-circuitpython-dht
smbus

